+++ crawler/src/main.rs
@@ -108,9 +108,8 @@ async fn run(config: Config, db: DB, shutdown: Controller<String>) -> Result<()>
 	spawn_in_span(shutdown.with_cancel(async move {
 		info!("Bootstraping the DHT with bootstrap nodes...");
 		let bootstraps = &config.libp2p.bootstraps;
-		match bootstrap_p2p_client.bootstrap_on_startup(bootstraps).await {
-			Ok(()) => info!("Bootstrap done."),
-			Err(e) => warn!("Bootstrap error: {e:?}."),
+		if let Err(error) = bootstrap_p2p_client.bootstrap_on_startup(bootstraps).await {
+			warn!("Bootstrap unsuccessful: {error:#}");
 		}
 	}));